Add support for setting link target as an option
This add support for setting the linkTarget option to specify where a link should open.

Default behavior is as today and will open in same target as current (_self).

Setting linkTarget to _blank wil open in a new window or tab or you can use a framename of your choice